refactor: rearrange the file
