+++ data/member/insert_new_member.py
@@ -1,4 +1,4 @@
-from data.create_connection_pool import create_connection_pool
+from utils.create_connection_pool import create_connection_pool
 
 db_pool = create_connection_pool("members_pool", "trip_members")
